requirements.txt
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-filter==25.1
drf-spectacular==0.28.0
django-anymail==13.0
httpx==0.28.1
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==6.0.3
django-celery-beat==2.9.0
django-timezone-field==7.2.1
flower==2.0.1
gunicorn==25.3.0
humanize==4.15.0
kombu==5.6.2
packaging==26.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2026.1.post1
redis==7.3.0
six==1.17.0
sqlparse==0.5.5
tornado==6.5.5
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
